Add dependency: mutils
