+++ matrix_sync/client/receiver.py
@@ -1,11 +1,11 @@
 # thread MatrixReceiver
 import asyncio
-from ..logger.get_logger import console_logger
 import matrix_sync.plg_globals as plg_globals
 
 from . import *
 from ..utils.token import getToken
-from ..utils import tr
+from ..utils.get_logger import console_logger
+from mutils import tr
 from ..event import *
 from nio import MatrixRoom, RoomMessageText, SyncError, UploadFilterError
